+++ lib/kitchen/directions/bake_free_response/v1.rb
@@ -12,7 +12,7 @@ module Kitchen::Directions::BakeFreeResponse
         free_response_questions = page.free_response
         next if free_response_questions.none?
 
-        free_response_questions.search('h3').trash
+        free_response_questions.search('h3')&.first&.trash
         title = Kitchen::Directions::EocSectionTitleLinkSnippet.v1(page: page)
         free_response_questions.each do |free_response_question|
           free_response_question.prepend(child: title)