+++ lib/kitchen/directions/bake_learning_objectives.rb
@@ -2,17 +2,22 @@
 
 module Kitchen
   module Directions
-    # Adds learning objectives header to abstracts
-    module BakePageAbstracts
+    module BakeLearningObjectives
       def self.v1(chapter:)
         chapter.abstracts.each do |abstract|
           abstract.prepend(child: "<h3 data-type='title'>#{I18n.t(:learning_objectives)}</h3>")
         end
       end
 
-      def self.v2(chapter:)
-        chapter.abstracts.each do |abstract|
-          abstract.prepend(child: "<h3 data-type='title'>#{I18n.t(:learning_objectives)}</h3>")
+      def self.v2(chapter:, add_title: true)
+        learning_objectives =
+          chapter.abstracts.any? ? chapter.abstracts : chapter.learning_objectives
+
+        learning_objectives.each do |abstract|
+          if add_title
+            abstract.prepend(child: "<h3 data-type='title'>#{I18n.t(:learning_objectives)}</h3>")
+          end
+
           ul = abstract.first!('ul')
           ul.add_class('os-abstract')
           ul.search('li').each_with_index do |li, index|