+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Gets rid of extraneous titles in `BakeAutoTitledNotes` when subtitles are off (minor)
 * Adds `BakeAutotitledExercise` direction and the option to `bake_unclassified_exercises` within `BakeAutotitledNotes`
 * Adds optional numbering for `BakeReferences.v1` (minor)
+* Patches`BakeNumberedNotes.v3` to suppress solutions outside examples when suppress_solutions is true (minor)
 
 ## [11.2.0] - 2021-09-10